Implement the JACK API method jack_set_client_registration_callback
The JACK API method jack_set_client_registration_callback is used
by a client to monitor connections and disconnections of other clients.

In turn, we will call a user supplied JS callback function with the
client name, and a boolean (connected).

Within the JACK callback context it is unsafe to access V8 internals,
so uv_async_send is used to notify a "safe" context where the callback
is actually invoked. libuv may coalesce calls to the uv_async_cb so
a linked list of information is maintained.